HeapAddress(usize) => HeapAddress(u64)
From what I understand, the most I can get into memory is going to be
usize, so theoretically usize should be fine.  However, right now the
algorithm for getting a new address isn't great and is going to
eventually start giving erroneous results once you hit max value (or
maybe panic).

u64 gives some more space.  Might bump it up to u64 * u64.  However,
ultimately there needs to be a garbage collection step that will free up
address space which needs some way to be efficiently walked through

+++ src/runtime.rs
@@ -2,7 +2,7 @@
 use purple::data::Func;
 
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq)]
-pub struct HeapAddress(pub usize);
+pub struct HeapAddress(pub u64);
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum RuntimeData {
